+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/client/ByProjectKeyBusinessUnitsRequestBuilder.java
@@ -48,6 +48,18 @@ public class ByProjectKeyBusinessUnitsRequestBuilder {
         return new ByProjectKeyBusinessUnitsByIDRequestBuilder(apiHttpClient, projectKey, ID);
     }
 
+    public ByProjectKeyBusinessUnitsKeyByKeyAssociatesByAssociateIdRequestBuilder keyWithKeyValueAssociatesWithAssociateIdValue(
+            String key, String associateId) {
+        return new ByProjectKeyBusinessUnitsKeyByKeyAssociatesByAssociateIdRequestBuilder(apiHttpClient, projectKey,
+            key, associateId);
+    }
+
+    public ByProjectKeyBusinessUnitsByBusinessUnitIdAssociatesByAssociateIdRequestBuilder withBusinessUnitIdValueAssociatesWithAssociateIdValue(
+            String businessUnitId, String associateId) {
+        return new ByProjectKeyBusinessUnitsByBusinessUnitIdAssociatesByAssociateIdRequestBuilder(apiHttpClient,
+            projectKey, businessUnitId, associateId);
+    }
+
     public ByProjectKeyBusinessUnitsSearchRequestBuilder search() {
         return new ByProjectKeyBusinessUnitsSearchRequestBuilder(apiHttpClient, projectKey);
     }